+++ app_utils/esi.py
@@ -3,23 +3,26 @@
 import datetime as dt
 import logging
 import random
-from time import sleep
+import warnings
+from contextlib import contextmanager
 from typing import Optional
 
-import requests
+from bravado.exception import HTTPError
+from celery import Task
 
 from django.utils.timezone import now
+from esi.clients import EsiClientProvider
 
-from app_utils.logging import LoggerAddTag
-
-from . import __title__, __version__
-from ._app_settings import (
+from app_utils import __title__, __version__
+from app_utils._app_settings import (
     APPUTILS_ESI_DAILY_DOWNTIME_END,
     APPUTILS_ESI_DAILY_DOWNTIME_START,
-    APPUTILS_ESI_ERROR_LIMIT_THRESHOLD,
 )
+from app_utils.logging import LoggerAddTag
 
 logger = LoggerAddTag(logging.getLogger(__name__), __title__)
+
+_esi = EsiClientProvider(ua_appname="allianceauth-app-utils", ua_version=__version__)
 
 
 class EsiStatusException(Exception):
@@ -48,10 +51,19 @@ class EsiDailyDowntime(EsiOffline):
 
 
 class EsiErrorLimitExceeded(EsiStatusException):
-    """ESI error limit exceeded error."""
+    """ESI error limit exceeded error.
+
+    DEPRECATED: This feature no longer works due to recent changes
+    of the status endpoint. This class will be removed in future versions.
+    """
 
     def __init__(self, retry_in: float = 60) -> None:
         """:meta private:"""
+        warnings.warn(
+            "EsiErrorLimitExceeded is deprecated and will be removed in future versions.",
+            category=DeprecationWarning,
+            stacklevel=2,
+        )
         super().__init__("The ESI error limit has been exceeded.")
         self._retry_in = float(retry_in)
 
@@ -62,9 +74,9 @@ class EsiErrorLimitExceeded(EsiStatusException):
 
 
 class EsiStatus:
-    """Current status of ESI (immutable)."""
-
-    MAX_JITTER = 20
+    """Represents the current online status of ESI
+    and can report whether the current time is within th the planned daily downtime.
+    """
 
     def __init__(
         self,
@@ -75,17 +87,23 @@ class EsiStatus:
     ) -> None:
         self._is_online = bool(is_online)
         self._is_daily_downtime = is_daily_downtime
-        if error_limit_remain is None or error_limit_reset is None:
-            self._error_limit_remain = None
-            self._error_limit_reset = None
-        else:
-            self._error_limit_remain = int(error_limit_remain)
-            self._error_limit_reset = int(error_limit_reset)
+        if error_limit_remain:
+            warnings.warn(
+                "error_limit_remain is deprecated and will be removed in future versions.",
+                category=DeprecationWarning,
+                stacklevel=2,
+            )
+        if error_limit_reset:
+            warnings.warn(
+                "error_limit_reset is deprecated and will be removed in future versions.",
+                category=DeprecationWarning,
+                stacklevel=2,
+            )
 
     @property
     def is_ok(self) -> bool:
-        """True if ESI is online and below error limit, else False."""
-        return self.is_online and not self.is_error_limit_exceeded
+        """True if ESI was online when this object was created."""
+        return self.is_online
 
     @property
     def is_daily_downtime(self) -> bool:
@@ -94,38 +112,51 @@ class EsiStatus:
 
     @property
     def is_online(self) -> bool:
-        """True if ESI is online, else False."""
+        """True if ESI was online when this object was created."""
         return self._is_online
 
     @property
     def error_limit_remain(self) -> Optional[int]:
-        """Amount of remaining errors in current window."""
-        return self._error_limit_remain
+        """Amount of remaining errors in current window.
+
+        DEPRECATED: This feature no longer works due to recent changes
+        of the status endpoint. This property will be removed in future versions.
+        """
+        warnings.warn(
+            "error_limit_remain is deprecated and will be removed in future versions.",
+            category=DeprecationWarning,
+            stacklevel=2,
+        )
+        return None
 
     @property
     def error_limit_reset(self) -> Optional[int]:
-        """Seconds until current error window resets."""
-        return self._error_limit_reset
+        """Seconds until current error window resets.
+
+        DEPRECATED: This feature no longer works due to recent changes
+        of the status endpoint. This property will be removed in future versions.
+        """
+        warnings.warn(
+            "error_limit_reset is deprecated and will be removed in future versions.",
+            category=DeprecationWarning,
+            stacklevel=2,
+        )
+        return None
 
     @property
     def is_error_limit_exceeded(self) -> bool:
         """True when remain is below the threshold, else False.
+        Will also return False if remain/reset are not defined.
 
-        Will also return False if remain/reset are not defined
+        DEPRECATED: This feature no longer works due to recent changes
+        of the status endpoint. This property will be removed in future versions.
         """
-        return bool(
-            self.error_limit_remain
-            and self.error_limit_reset
-            and self.error_limit_remain <= APPUTILS_ESI_ERROR_LIMIT_THRESHOLD
+        warnings.warn(
+            "is_error_limit_exceeded is deprecated and will be removed in future versions.",
+            category=DeprecationWarning,
+            stacklevel=2,
         )
-
-    def error_limit_reset_w_jitter(self, max_jitter: Optional[int] = None) -> int:
-        """Calc seconds to retry in order to reach next error window incl. jitter."""
-        if self.error_limit_reset is None:
-            return 0
-        if not max_jitter or max_jitter < 1:
-            max_jitter = self.MAX_JITTER
-        return self.error_limit_reset + int(random.uniform(1, max_jitter))
+        return False
 
     def raise_for_status(self):
         """Raise an exception if ESI if offline or the error limit is exceeded."""
@@ -133,12 +164,10 @@ class EsiStatus:
             if self.is_daily_downtime:
                 raise EsiDailyDowntime()
             raise EsiOffline()
-        if self.is_error_limit_exceeded:
-            raise EsiErrorLimitExceeded(retry_in=self.error_limit_reset_w_jitter())
 
 
 def fetch_esi_status(ignore_daily_downtime: bool = False) -> EsiStatus:
-    """Determine the current ESI status.
+    """Determine and return the current ESI status.
 
     Args:
         ignore_daily_downtime: When True will always make a request to ESI \
@@ -147,36 +176,19 @@ def fetch_esi_status(ignore_daily_downtime: bool = False) -> EsiStatus:
     is_daily_downtime = _is_daily_downtime()
     if not ignore_daily_downtime and is_daily_downtime:
         return EsiStatus(is_online=False, is_daily_downtime=True)
+
     try:
-        response = _request_esi_status()
-    except (requests.exceptions.Timeout, requests.exceptions.ConnectionError):
+        status = _esi.client.Status.get_status().result(retries=1)
+    except ConnectionError:
         logger.warning("Network error when trying to call ESI", exc_info=True)
         return EsiStatus(is_online=False, is_daily_downtime=is_daily_downtime)
-    if not response.ok:
-        is_online = False
-    else:
-        try:
-            is_online = not response.json().get("vip")
-        except ValueError:
-            is_online = False
-    try:
-        remain = int(response.headers.get("X-Esi-Error-Limit-Remain"))  # type: ignore
-        reset = int(response.headers.get("X-Esi-Error-Limit-Reset"))  # type: ignore
-    except TypeError:
-        logger.warning("Failed to parse HTTP headers: %s", response.headers)
-        return EsiStatus(is_online=is_online, is_daily_downtime=is_daily_downtime)
-    logger.debug(
-        "ESI status: is_online: %s, error_limit_remain = %s, error_limit_reset = %s",
-        is_online,
-        remain,
-        reset,
-    )
-    return EsiStatus(
-        is_online=is_online,
-        error_limit_remain=remain,
-        error_limit_reset=reset,
-        is_daily_downtime=is_daily_downtime,
-    )
+    except HTTPError:  # Will usually return http error 502 when offline
+        logger.warning("HTTP error when trying to call ESI", exc_info=True)
+        return EsiStatus(is_online=False, is_daily_downtime=is_daily_downtime)
+
+    is_online = status.get("vip") is None
+    logger.debug("ESI status: is_online: %s", is_online)
+    return EsiStatus(is_online=is_online, is_daily_downtime=is_daily_downtime)
 
 
 def _is_daily_downtime() -> bool:
@@ -198,47 +210,13 @@ def _convert_float_hours(hours_float: float) -> tuple:
     return hours, minutes
 
 
-def _request_esi_status() -> requests.Response:
-    """Fetch current status from ESI. Retry on common HTTP errors."""
-    max_retries = 3
-    retry_count = 0
-    while True:
-        response = requests.get(
-            "https://esi.evetech.net/latest/status/",
-            timeout=(5, 30),
-            headers={"User-Agent": f"{__package__};{__version__}"},
-        )
-        if response.status_code not in {
-            502,  # HTTPBadGateway
-            503,  # HTTPServiceUnavailable
-            504,  # HTTPGatewayTimeout
-        }:
-            break
-
-        retry_count += 1
-        if retry_count > max_retries:
-            break
-
-        logger.warning(
-            "HTTP status code %s - Try %s/%s",
-            response.status_code,
-            retry_count,
-            max_retries,
-        )
-
-        wait_secs = 0.1 * (random.uniform(2, 4) ** (retry_count - 1))
-        sleep(wait_secs)
-
-    return response
-
-
-def retry_task_if_esi_is_down(self):
-    """Retry current celery task if ESI is not online or error threshold is exceeded.
+def retry_task_if_esi_is_down(task: Task):
+    """Retry current celery task if ESI is not online.
 
     This function has to be called from inside a celery task!
 
     Args:
-        self: Current celery task from `@shared_task(bind=True)`
+        task: Current celery task from `@shared_task(bind=True)`
     """
     try:
         fetch_esi_status().raise_for_status()
@@ -247,9 +225,51 @@ def retry_task_if_esi_is_down(self):
         logger.warning(
             "ESI appears to be offline. Trying again in %d seconds.", countdown
         )
-        raise self.retry(countdown=countdown) from ex
-    except EsiErrorLimitExceeded as ex:
-        logger.warning(
-            "ESI error limit threshold reached. Trying again in %s seconds", ex.retry_in
-        )
-        raise self.retry(countdown=ex.retry_in) from ex
+        raise task.retry(countdown=countdown) from ex
+
+
+@contextmanager
+def retry_task_on_esi_error_and_offline(task: Task, info: str):
+    """Context manager that retries a task when the error error limit has been exceeded
+    or when ESI appears to be offline.
+
+    Args:
+    - task: current celery task
+    - info: text describing what is being retried
+
+    Example:
+
+    .. code-block:: python
+
+        from app_utils.esi import retry_task_on_esi_error_and_offline
+
+        @shared_task(bind=True)
+        def my_task(self):
+             with retry_task_on_esi_error_and_offline(self, "my_task"):
+                # work that might trigger an HTTPError
+
+    '''
+    """
+    try:
+        yield
+    except HTTPError as ex:
+        backoff_jitter = int(random.uniform(2, 4) ** task.request.retries)
+        if ex.status_code == 420:  # ESI error limit exceeded
+            countdown = 60 + backoff_jitter
+            logger.warning(
+                "%s: ESI error limit exceeded. Trying again in %s seconds",
+                info,
+                countdown,
+            )
+            raise task.retry(countdown=countdown) from ex
+
+        if ex.status_code in {502, 503}:  # ESI offline
+            countdown = (15 + backoff_jitter) * 60
+            logger.warning(
+                "%s: ESI appears to be offline. Trying again in %d seconds",
+                info,
+                countdown,
+            )
+            raise task.retry(countdown=countdown) from ex
+
+        raise ex