+++ app_utils/esi_testing.py
@@ -87,6 +87,7 @@ def build_http_error(http_code: int, text: Optional[str] = None) -> HTTPError:
         401: HTTPUnauthorized,
         403: HTTPForbidden,
         404: HTTPNotFound,
+        420: HTTPError,
         500: HTTPInternalServerError,
         502: HTTPBadGateway,
         503: HTTPServiceUnavailable,